move list_all_topics_actor
