+++ ydb/core/persqueue/events/ya.make
@@ -9,6 +9,7 @@ PEERDIR(
     ydb/core/base
     ydb/core/keyvalue
     ydb/core/protos
+    ydb/core/persqueue/public/counters
     ydb/core/tablet
     ydb/public/api/protos
     ydb/library/persqueue/topic_parser
