+++ ydb/core/persqueue/pqtablet/cache/ya.make
@@ -9,6 +9,7 @@ SRCS(
 PEERDIR(
     ydb/core/base
     ydb/core/keyvalue
+    ydb/core/persqueue/pqtablet/blob
     ydb/library/logger
 )
