+++ ydb/core/persqueue/pqtablet/partition/mirrorer.h
@@ -5,7 +5,7 @@
 #include <ydb/library/actors/core/hfunc.h>
 #include <ydb/library/actors/core/log.h>
 #include <ydb/core/base/appdata.h>
-#include <ydb/core/persqueue/public/percentile_counter.h>
+#include <ydb/core/persqueue/public/counters/percentile_counter.h>
 #include <ydb/core/protos/counters_pq.pb.h>
 #include <ydb/core/protos/pqconfig.pb.h>
 #include <ydb/public/lib/base/msgbus.h>
