+++ ydb/core/persqueue/pqtablet/ya.make
@@ -13,6 +13,7 @@ SRCS(
 PEERDIR(
     ydb/core/persqueue/pqtablet/common
     ydb/core/persqueue/common/proxy
+    ydb/core/persqueue/public/counters
     ydb/core/persqueue/pqtablet/cache
     ydb/core/persqueue/pqtablet/partition
 )
