+++ ydb/core/persqueue/public/counters/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    percentile_counter.cpp
+)
+
+PEERDIR(
+    library/cpp/monlib/dynamic_counters
+    ydb/library/persqueue/topic_parser
+)
+
+END()
