+++ ydb/core/persqueue/public/list_topics/list_all_topics_actor.h
@@ -1,7 +1,7 @@
 #include <util/generic/maybe.h>
 #include <ydb/library/actors/core/actor.h>
 
-namespace NKikimr::NPersQueue {
+namespace NKikimr::NPQ {
 
 NActors::IActor* MakeListAllTopicsActor(const NActors::TActorId& respondTo, const TString& databasePath, const TString& token,
                                         bool recursive, const TString& startFrom = {}, const TMaybe<ui64>& limit = Nothing());
