+++ ydb/core/persqueue/public/ya.make
@@ -1,7 +1,6 @@
 LIBRARY()
 
 SRCS(
-    percentile_counter.cpp
     pq_database.cpp
     pq_rl_helpers.cpp
     utils.cpp
@@ -15,6 +14,8 @@ PEERDIR(
 END()
 
 RECURSE(
+    counters
     cluster_tracker
     fetcher
+    list_topics
 )
