+++ ydb/core/persqueue/ut/counters_ut.cpp
@@ -4,12 +4,14 @@
 #include <ydb/core/base/tablet_pipecache.h>
 #include <ydb/core/mon/mon.h>
 #include <ydb/core/persqueue/ut/common/pq_ut_common.h>
-#include <ydb/core/persqueue/public/percentile_counter.h>
+#include <ydb/core/persqueue/public/counters/percentile_counter.h>
 #include <ydb/core/persqueue/pqtablet/partition/partition.h>
 #include <ydb/core/sys_view/service/sysview_service.h>
 #include <ydb/core/testlib/fake_scheme_shard.h>
 #include <ydb/core/tx/scheme_cache/scheme_cache.h>
 #include <ydb/core/keyvalue/keyvalue_events.h>
+#include <ydb/public/api/protos/draft/persqueue_error_codes.pb.h>
+
 namespace NKikimr::NPQ {
 
 namespace {
