+++ ydb/services/datastreams/datastreams_proxy.cpp
@@ -17,7 +17,7 @@
 #include <ydb/services/lib/actors/pq_schema_actor.h>
 #include <ydb/services/lib/sharding/sharding.h>
 #include <ydb/services/persqueue_v1/actors/persqueue_utils.h>
-#include <ydb/core/persqueue/list_all_topics_actor.h>
+#include <ydb/core/persqueue/public/list_topics/list_all_topics_actor.h>
 
 #include <util/folder/path.h>
 
@@ -953,7 +953,7 @@ namespace NKikimr::NDataStreams::V1 {
             Request_->ReplyWithYdbStatus(Ydb::StatusIds::BAD_REQUEST);
             return Die(ctx);
         }
-        ctx.Register(NPersQueue::MakeListAllTopicsActor(SelfId(), Request_->GetDatabaseName().GetOrElse(""),
+        ctx.Register(NPQ::MakeListAllTopicsActor(SelfId(), Request_->GetDatabaseName().GetOrElse(""),
                                                         this->Request_->GetSerializedToken(), GetProtoRequest()->recurse(),
                                                         GetProtoRequest()->exclusive_start_stream_name(), limit));
         Become(&TListStreamsActor::StateWork);
