+++ ydb/services/deprecated/persqueue_v0/ya.make
@@ -20,6 +20,7 @@ PEERDIR(
     ydb/core/mind/address_classification
     ydb/core/persqueue
     ydb/core/persqueue/events
+    ydb/core/persqueue/public/counters
     ydb/core/persqueue/writer
     ydb/core/protos
     ydb/library/aclib
