+++ ydb/services/persqueue_v1/actors/events.h
@@ -6,7 +6,7 @@
 #include <ydb/core/grpc_services/rpc_calls_topic.h>
 #include <ydb/core/protos/pqconfig.pb.h>
 #include <ydb/core/persqueue/public/key.h>
-#include <ydb/core/persqueue/public/percentile_counter.h>
+#include <ydb/core/persqueue/public/counters/percentile_counter.h>
 #include <ydb/core/tx/scheme_board/events.h>
 
 #include <ydb/public/api/protos/persqueue_error_codes_v1.pb.h>
